Split models to 5 sections and school service

+++ IISportSchool/Models/Persistence/EFRepositories/EFChildrenRepository.cs
@@ -24,8 +24,11 @@ namespace IISportSchool.Models
 
         public Children Delete(Children child)
         {
-            _context.Childrens.Remove(child);
-            _context.SaveChanges();
+            if(child != null)
+            {
+                _context.Childrens.Remove(child);
+                _context.SaveChanges();
+            }
             return child;
         }
